routes structure changed, auth route implemented
