+++ routes/auth.js
@@ -1,0 +1,19 @@
+const router = require('express').Router();
+const { celebrate, Joi } = require('celebrate');
+const { login, createUser } = require('../controllers/auth');
+
+router.post('/signin', celebrate({
+  body: Joi.object().keys({
+    email: Joi.string().required().email(),
+    password: Joi.string().required().min(2).max(30),
+  }),
+}), login);
+router.post('/signup', celebrate({
+  body: Joi.object().keys({
+    email: Joi.string().required().email(),
+    password: Joi.string().required().min(2).max(30),
+    name: Joi.string().required().min(2).max(30),
+  }),
+}), createUser);
+
+module.exports = router;
